refac: split types into organized files
