+++ src/types/Graph.ts
