+++ src/types/geometry.ts
@@ -1,0 +1,10 @@
+export interface Point {
+    x: number;
+    y: number;
+}
+
+export interface ViewportTransform {
+    x: number;
+    y: number;
+    scale: number;
+}
