+++ src/types/modes.ts
@@ -1,0 +1,6 @@
+export interface Mode {
+    name: string;
+    handleDragging(e: MouseEvent): void;
+    handleClick(e: MouseEvent): void;
+    handleMouseMove(e: MouseEvent): void;
+}
